% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/target_skeleton.R
\name{target_skeleton}
\alias{target_skeleton}
\title{Create target skeleton directory}
\usage{
target_skeleton(scratch_path = "", include_examples = TRUE)
}
\arguments{
\item{scratch_path}{If set, tells targets to use the specified directory for its _targets objects}

\item{include_examples}{If TRUE, includes some example commands in files}
}
\description{
This creates necessary target files and templates in the current directory. You can optionally set a different directory (such as a dropbox folder) to hold the _targets directory.
}
